mooved hook related stuff into separate file
